+++ src/main/java/com/groom/orbit/schedule/application/ScheduleCommandService.java
@@ -1,4 +1,4 @@
-package com.groom.orbit.schedule.app;
+package com.groom.orbit.schedule.application;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -6,9 +6,9 @@ import org.springframework.transaction.annotation.Transactional;
 import com.groom.orbit.common.dto.CommonSuccessDto;
 import com.groom.orbit.member.member.app.MemberQueryService;
 import com.groom.orbit.member.member.dao.jpa.entity.Member;
-import com.groom.orbit.schedule.app.dto.ScheduleRequestDto;
-import com.groom.orbit.schedule.dao.Schedule;
-import com.groom.orbit.schedule.dao.ScheduleRepository;
+import com.groom.orbit.schedule.application.dto.ScheduleRequestDto;
+import com.groom.orbit.schedule.repository.jpa.ScheduleRepository;
+import com.groom.orbit.schedule.repository.jpa.entity.Schedule;
 
 import lombok.RequiredArgsConstructor;
