+++ src/main/java/com/groom/orbit/schedule/controller/CalendarQueryController.java
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.groom.orbit.common.annotation.AuthMember;
 import com.groom.orbit.common.dto.ResponseDto;
-import com.groom.orbit.schedule.app.ScheduleQueryService;
-import com.groom.orbit.schedule.app.dto.GetCalendarResponseDto;
+import com.groom.orbit.schedule.application.ScheduleQueryService;
+import com.groom.orbit.schedule.application.dto.GetCalendarResponseDto;
 
 import lombok.RequiredArgsConstructor;
