+++ src/main/java/com/groom/orbit/schedule/controller/ScheduleCreateController.java
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.RestController;
 import com.groom.orbit.common.annotation.AuthMember;
 import com.groom.orbit.common.dto.CommonSuccessDto;
 import com.groom.orbit.common.dto.ResponseDto;
-import com.groom.orbit.schedule.app.ScheduleCommandService;
-import com.groom.orbit.schedule.app.dto.ScheduleRequestDto;
+import com.groom.orbit.schedule.application.ScheduleCommandService;
+import com.groom.orbit.schedule.application.dto.ScheduleRequestDto;
 
 import lombok.RequiredArgsConstructor;
