+++ src/main/java/com/groom/orbit/schedule/controller/ScheduleUpdateController.java
@@ -5,8 +5,8 @@ import org.springframework.web.bind.annotation.*;
 import com.groom.orbit.common.annotation.AuthMember;
 import com.groom.orbit.common.dto.CommonSuccessDto;
 import com.groom.orbit.common.dto.ResponseDto;
-import com.groom.orbit.schedule.app.ScheduleCommandService;
-import com.groom.orbit.schedule.app.dto.ScheduleRequestDto;
+import com.groom.orbit.schedule.application.ScheduleCommandService;
+import com.groom.orbit.schedule.application.dto.ScheduleRequestDto;
 
 import lombok.RequiredArgsConstructor;
