+++ src/main/java/com/groom/orbit/schedule/repository/jpa/entity/Schedule.java
@@ -1,16 +1,8 @@
-package com.groom.orbit.schedule.dao;
+package com.groom.orbit.schedule.repository.jpa.entity;
 
 import java.time.LocalDate;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
 
 import org.hibernate.annotations.DynamicUpdate;
